% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy.R
\name{canopy}
\alias{canopy}
\title{Calculate canopy closure from S.W.T.S. survey 2020}
\usage{
canopy(myVector)
}
\arguments{
\item{myVector}{A vector contains canopy opennes data from densiometer.}
}
\value{
Mean of the canopy closure for a spatial replicate.
}
\description{
A function to calculate mean of canopy closure from each spatial replicate.
}
