% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearPoint.R
\name{clearPoint}
\alias{clearPoint}
\title{Remove nearby points within specific distance in meters}
\usage{
clearPoint(dataFrame, Xcol, Ycol, UTMZone, distLength)
}
\arguments{
\item{dataFrame}{A data-frame contains at least "X" and "Y" in UTM projection.}

\item{Xcol}{A quoted name of column that consists X coordinates.}

\item{Ycol}{A quoted name of column that consists Y coordinates.}

\item{UTMZone}{A UTM projection of the target file.}

\item{distLength}{A numeric value of a distance}
}
\value{
Spatial object data-frame.
}
\description{
A function to remove points within a distance.
}
