% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copyID_2.R
\name{copyID_2}
\alias{copyID_2}
\title{Copy attributes from other spatial points data-frame}
\usage{
copyID_2(points1, points2)
}
\arguments{
\item{points1}{SpatialPointsDataframe consist of X and Y that copy attributes from points2.}

\item{points2}{SpatialPointsDataframe where the attributes available to be copied.}
}
\value{
Similar data with points1 with additional ID's copied from nearby points2.
}
\description{
A function to copy ID's from other nearby SpatialPointsDataframe.
}
