% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist2area.R
\name{dist2area}
\alias{dist2area}
\title{dist2area}
\usage{
dist2area(inDist, outDist)
}
\arguments{
\item{inDist}{a raster layer of distance within the area, the values increase from border to the centre of the polygone.}

\item{outDist}{a raster layer of distance outside the area, values are negative from -1 close to the edge to -inf further away from the border.}
}
\value{
Distance to area (positive values inside, negative outside).
}
\description{
A function to combine and stitch two distance raster together.
}
