% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist3D.R
\name{dist3D}
\alias{dist3D}
\title{Calculate the distance in three dimensions if X, Y, and Z coordinates are available.}
\usage{
dist3D(dataFrame, Xcol, Ycol, elevData, repLength, distType = "3D")
}
\arguments{
\item{dataFrame}{A data frame containing information, including X and Y coordinates in UTM projection.}

\item{Xcol}{A quoted column name representing X coordinates within the dataFrame.}

\item{Ycol}{A quoted column name representing Y coordinates within the dataFrame.}

\item{elevData}{Elevation data in meters in "SpatRaster" format, utilized for extracting Z values.}

\item{repLength}{A predefined replication length specified in meters.}

\item{distType}{Specify whether to calculate distance in 3D or 2D (Default is 3D).}
}
\value{
New columns specifying the distance and the corresponding replicates to which the observations belong.
}
\description{
A function designed to compute the 3D distance and subsequently categorize observations according to a predefined replication length.
}
