% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm2spatial.R
\name{dm2spatial}
\alias{dm2spatial}
\title{Transform the detection matrix into an sf object.}
\usage{
dm2spatial(detectMatrix, spProject)
}
\arguments{
\item{detectMatrix}{A matrix containing species detection/non-detection, generated from 'speciesDM'.}

\item{spProject}{Spatial object with its projection to be duplicated.}
}
\value{
An sf data frame containing species detection/non-detection.
}
\description{
A function to transform the detection matrix into an sf object.
}
