% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_point.R
\name{grid_point}
\alias{grid_point}
\title{Calculate proportion of points over SpatialPointsDataframe.}
\usage{
grid_point(
  spPoints,
  subGrids,
  whichCol,
  whichSp,
  as_Raster = TRUE,
  rasRes = 1000
)
}
\arguments{
\item{spPoints}{A SpatialPointsDataframe in UTM projection.}

\item{subGrids}{Grid cells within the landscape of interest, each grid cell can be identified with id column, also in UTM.}

\item{whichCol}{Specific field or column where the features exist.}

\item{whichSp}{Specific type or category to be selected as presence/absence.}

\item{as_Raster}{if TRUE will create raster data.}

\item{rasRes}{Raster resolution (default 1000 mtr).}
}
\value{
Proportion of points.
}
\description{
A function to calculate proportion of points over SpatialPointsDataframe.
}
