% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_sf_fragstat.R
\name{grid_sf_fragstat}
\alias{grid_sf_fragstat}
\title{Calculate landscape matrices within gridcells}
\usage{
grid_sf_fragstat(
  landCover,
  landLevel = "landscape",
  lClass = FALSE,
  subGrids,
  lsmFunc,
  as_Raster = TRUE,
  rasRes = 1000
)
}
\arguments{
\item{landCover}{Raster in terraVect format consists of land cover/use classes in UTM projection.}

\item{landLevel}{Specify matrix level(only landscape and class level available)}

\item{lClass}{Specify the class for extraction (e.g., 1 for forest, 2 for cropland)}

\item{subGrids}{Grid cells within the landscape of interest, each grid cell can be identified with id column, also in UTM.}

\item{lsmFunc}{A function from landscapemetrices package.}

\item{as_Raster}{TRUE, output will be converted to raster (TRUE).}

\item{rasRes}{If as_Raster = TRUE, then specify the resolution. Default is 1000m.}
}
\value{
Gridcells in raster with 1 km resolution or polygon with values from landscapesmetrices function.
}
\description{
A function to calculate landscape matrices (fragmentation statistics) within gridcells.
}
