% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line2points.R
\name{line2points}
\alias{line2points}
\title{Transforming a spatial lines data frame into a spatial points data frame.}
\usage{
line2points(spLineDF, minDist)
}
\arguments{
\item{spLineDF}{spatial lines data frame representing a track observed during a transect survey.}

\item{minDist}{The minimum distance between points, measured in meters.}
}
\value{
A spatial points data frame derived from the transformation of a spatial lines data frame.
}
\description{
A function designed to convert a line track into an ordered set of points along that track.
}
