% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line2points_2.R
\name{line2points_2}
\alias{line2points_2}
\title{Convert SpatialLinesDataframe into SpatialPointsDataframe}
\usage{
line2points_2(spLineDF, minDist)
}
\arguments{
\item{spLineDF}{A SpatialLinesDataFrame of a track from a transect survey.}

\item{minDist}{Minimum distance between points.}
}
\value{
A SpatialPointsDataFrame from the line track.
}
\description{
A function to turn a line track into an ordered points along that line track
}
