% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineDens_grid.R
\name{lineDens_grid}
\alias{lineDens_grid}
\title{Compute line density for individual grid cells and convert the results into a raster format.}
\usage{
lineDens_grid(sfData, subGrids, as_Raster = TRUE, rasRes = 1000)
}
\arguments{
\item{sfData}{An sf object of multiple lines in UTM projection.}

\item{subGrids}{Grid cells within the landscape of interest, each grid cell can be identified with id column, also in UTM.}

\item{as_Raster}{If the output converted to raster (TRUE).}

\item{rasRes}{If as_Raster = TRUE, then specify the resolution. Default is 1000m.}
}
\value{
Density of line in km/sqkm unit area in raster or polygon.
}
\description{
A function to compute line density for individual grid cells and convert the results into a raster format.
}
