% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean.R
\name{mean}
\alias{mean}
\title{Calculate mean}
\usage{
mean(myVector)
}
\arguments{
\item{myVector}{A vector contains numeric values.}
}
\value{
Mean of the vector.
}
\description{
A function to calculate mean from numeric vector.
}
