% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanPoint.R
\name{meanPoint}
\alias{meanPoint}
\title{Take a mean value from a set of coordinates in sequence}
\usage{
meanPoint(dataFrame, nPoint)
}
\arguments{
\item{dataFrame}{A spatial points data-frame.}

\item{nPoint}{A number of point to average.}
}
\value{
A data-frame contains the remainder and the division value.
}
\description{
A function clear multi-points downloaded from GPS (Global Positioning System).
}
