% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modal.R
\name{modal}
\alias{modal}
\title{Calculate the most common character values}
\usage{
modal(myVector)
}
\arguments{
\item{myVector}{A vector contains character values.}
}
\value{
the most common values from a vector of elements.
}
\description{
A function to measure the central tendency of the data (the most common characteristics of the sample).
}
