% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myCanopy.R
\name{myCanopy}
\alias{myCanopy}
\title{Calculating canopy closure from the Sumatran-wide Tiger Survey (SWTS) 2018-2020}
\usage{
myCanopy(myVector)
}
\arguments{
\item{myVector}{A vector containing canopy openness data obtained from densiometer measurements in the field.}
}
\value{
The mean of the canopy closure values within a spatial replicate..
}
\description{
This function is designed to calculate the mean canopy closure from a set of values representing canopy closure in each spatial replicate.
}
