% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myMean.R
\name{myMean}
\alias{myMean}
\title{Compute the mean from the given dataset.}
\usage{
myMean(myVector)
}
\arguments{
\item{myVector}{A vector containing numeric values.}
}
\value{
The mean of the given vector.
}
\description{
A function designed to calculate the mean from a numeric vector.
}
