% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myModal.R
\name{myModal}
\alias{myModal}
\title{Compute the most common character values, also known as the mode.}
\usage{
myModal(myVector)
}
\arguments{
\item{myVector}{A vector comprising character values.}
}
\value{
The most common values from a vector containing elements of strings.
}
\description{
A function designed to assess the central tendency of the data, identifying the most common characteristics within the sample.
}
