% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearID.R
\name{nearID}
\alias{nearID}
\title{nearID}
\usage{
nearID(points1, points2, joinAtr)
}
\arguments{
\item{points1}{Spatial points dataframe consist of X and Y that needs attribute from points2.}

\item{points2}{Spatial points dataframe where the attributes available to be copied.}

\item{joinAtr}{A vector of attribute (columns) to be copied.}
}
\value{
Similar data with wayPoints with additional Id copied from nearby track points.
}
\description{
A function to copy IDs from other nearby track points.
}
