% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDM.R
\name{plotDM}
\alias{plotDM}
\title{Plot detection matrix of a species}
\usage{
plotDM(DetMatrix, label = TRUE, Title = "Title")
}
\arguments{
\item{DetMatrix}{A detection matrix for selected species.}

\item{label}{Whether to use label of the replicates (default = TRUE).}

\item{Title}{The title of the plot.}
}
\value{
A plot of species detection matrix.
}
\description{
A function to plot detection matrix for selected species.
}
