% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasFragstat_grid.R
\name{rasFragstat_grid}
\alias{rasFragstat_grid}
\title{Calculate landscape fragmentation matrices within grid cells}
\usage{
rasFragstat_grid(
  landCover,
  landLevel = "landscape",
  lClass = FALSE,
  subGrids,
  lsmFunc,
  as_Raster = TRUE,
  rasRes = 1000
)
}
\arguments{
\item{landCover}{A raster in 'terraVect' format comprises land cover/use classes in UTM projection.}

\item{landLevel}{Specify the landscape matrix level (only 'landscape' and 'class' levels are available).}

\item{lClass}{Specify the class for extraction (e.g., 1 for forest, 2 for cropland).}

\item{subGrids}{Grid cells within the landscape of interest, each identified by an "id" column, also in UTM}

\item{lsmFunc}{A function or a set of functions imported from the 'landscapemetrics' package.}

\item{as_Raster}{If set to TRUE, the output will be converted to a raster.}

\item{rasRes}{If 'as_Raster' is set to TRUE, please specify the resolution. The default is 1000 meters.}
}
\value{
Grid cells in a raster with a 1 km resolution or polygons with values from the landscapemetrics function.
}
\description{
A function to calculate landscape matrices (fragmentation statistics) within grid cells.
}
