% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasStat_grid.R
\name{rasStat_grid}
\alias{rasStat_grid}
\title{Calculate statistic for each subgrids from a raster dataset.}
\usage{
rasStat_grid(rasterData, subGrids, myFunc, as_Raster = TRUE, rasRes = 1000)
}
\arguments{
\item{rasterData}{A terraVect data in UTM projection.}

\item{subGrids}{Grid cells within the landscape of interest, each grid cell can be identified with id column, also in UTM.}

\item{myFunc}{A statistical function such as mean, sum, max, etc to be carried out for each grid.}

\item{as_Raster}{If the output converted to raster (TRUE).}

\item{rasRes}{If as_Raster = TRUE, then specify the resolution. Default is 1000m.}
}
\value{
Gridcells in raster with 1 km resolution or polygons with raster values.
}
\description{
A function to calculate statistic for each subgrids from a raster dataset.
}
