% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorderPoint.R
\name{reorderPoint}
\alias{reorderPoint}
\title{Reorder point based on group characteristics}
\usage{
reorderPoint(pointsDF, startPoint, groupID)
}
\arguments{
\item{pointsDF}{A dataframe of points to be reordered, it should contain grouping Ids.}

\item{startPoint}{A specific row from pointsDF where the point starts.}

\item{groupID}{The Id that represents the group information.}
}
\value{
Similar data with wayPoints with additional Id copied from nearby track points.
}
\description{
A function to reorder point based on group information and distance among the points
}
