% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDM.R
\name{spatialDM}
\alias{spatialDM}
\title{Generate detection matrices specific to a particular species or entities along the transects.}
\usage{
spatialDM(
  speciesDF,
  sortID,
  Xcol,
  Ycol,
  whichCol,
  whichSp,
  samplingCov = FALSE,
  samplingFun = FALSE
)
}
\arguments{
\item{speciesDF}{A dataframe includes the species, location along the transects, and the replicate column generated from the dist3D function.}

\item{sortID}{A column used to order points based on a sequence, either time-based or ID-based.}

\item{Xcol}{A quoted column name representing X coordinates within the dataFrame.}

\item{Ycol}{A quoted column name representing Y coordinates within the dataFrame.}

\item{whichCol}{A column containing all the species for which to extract detection matrices .}

\item{whichSp}{A selected species or entity name within the "whichCol" column to be extracted as detection matrices.}

\item{samplingCov}{Sampling covariates recorded during the surveys that are to be extracted from each replicate (default is FALSE).}

\item{samplingFun}{A list of methods to handle samplingCov, specifically designed for SWTS survey, including only modal (for character), mean (for numeric), and canopy (to calculate canopy closure) functions, are currently available.}
}
\value{
A data frame containing the detection matrix for selected species or entities, along with their geographic coordinates and sampling covariates.
}
\description{
A function to Generate detection matrices specific to a particular species or entities along the transects..
}
