% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialDM_grid.R
\name{spatialDM_grid}
\alias{spatialDM_grid}
\title{Generate a spatial detection matrix (spatialDM) for a specific species or entities within specified grid cells.}
\usage{
spatialDM_grid(
  spData,
  sortID,
  repLength,
  gridCell,
  subgridCol,
  elevData,
  whichCol,
  whichSp,
  Xcol,
  Ycol,
  samplingCov = FALSE,
  samplingFun = FALSE
)
}
\arguments{
\item{spData}{A spatial points containing the recorded species or entities along the transect in sf format.}

\item{sortID}{A column name utilized to sequence the points, whether based on time or ID.}

\item{repLength}{User input specifying the desired length of each spatial replicate.}

\item{gridCell}{Spatial polygons or grid cells employed for the subdivision or intersection of the 'spData'.}

\item{subgridCol}{A column containing the subgrid ID used to partition the data based on grid cells.}

\item{elevData}{A raster layer containing elevation data in meters to calculate altitude (Z).}

\item{whichCol}{A column containing all the species for which to extract detection matrices.}

\item{whichSp}{A selected species or entity name within the "whichCol" column to be extracted as detection matrices.}

\item{Xcol}{A quoted column name representing X coordinates within the dataFrame.}

\item{Ycol}{A quoted column name representing Y coordinates within the dataFrame.}

\item{samplingCov}{Sampling covariates recorded during the surveys that are to be extracted from each replicate (default is FALSE).}

\item{samplingFun}{A list of methods to handle samplingCov, specifically designed for SWTS survey, including only modal (for character), mean (for numeric), and canopy (to calculate canopy closure) functions, are currently available.}
}
\value{
A data frame containing the detection matrix for selected species or entities, along with their geographic coordinates and sampling covariates from all gri9d cells.
}
\description{
A function to Generate a spatial detection matrix (spatialDM) for a specific species or entities within specified grid cells.
}
