% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesDM_grid.R
\name{speciesDM_grid}
\alias{speciesDM_grid}
\title{speciesDM_grid}
\usage{
speciesDM_grid(
  spData,
  sortID,
  repLength,
  gridCell,
  subgridCol,
  elevData,
  whichCol,
  whichSp,
  Xcol,
  Ycol,
  extractVars = FALSE
)
}
\arguments{
\item{spData}{A spatialPointsDataframe contain the occupancy data.}

\item{sortID}{A name of column that used to order the point based on a sequence (time based or ID based).}

\item{repLength}{An information about a desired length of each spatial replicate.}

\item{gridCell}{A spatialPolygonDataframe used to split/intersect the occu data.}

\item{subgridCol}{A column contain the subgrid id used to split the data based on gridcell.}

\item{elevData}{A raster layer contains elevation data in meter to calculate altitude (Z).}

\item{whichCol}{A column that contains all the species occurrence}

\item{whichSp}{A selected species name within the "whichCol" column to be extracted}

\item{Xcol}{A column that consists X coordinates.}

\item{Ycol}{A column that consists Y coordinates.}

\item{extractVars}{The variables (columns) to be extracted for each replicate. Default is FALSE}
}
\value{
A dataframe contain detection matrices from all gridcell.
}
\description{
A function to extract detection matrix for a species from all gridcells
}
