% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeFormat.R
\name{timeFormat}
\alias{timeFormat}
\title{timeFormat}
\usage{
timeFormat(myString, addTime)
}
\arguments{
\item{myString}{Information about date and time from tracklogs.}

\item{addTime}{A number of time to be added which happens when the timezone is not defined yet (e.g. Jakarta is UTC+7, so need to add 7 if the current time zone is UTC/GMT).}
}
\value{
a new string with DateTime format (as.Date or as.POSIXct).
}
\description{
A function to reformat date and time.
}
