\name{track}
\alias{track}
\docType{data}
\title{
Sample of track survey
}
\description{
A subset of track survey
}
\usage{data("track")}
\format{
  A data frame with 4706 observations on the following 4 variables.
  \describe{
    \item{\code{DateTime}}{a character vector}
    \item{\code{X}}{a numeric vector}
    \item{\code{Y}}{a numeric vector}
    \item{\code{Z}}{a numeric vector}
  }
}
\details{
A subset of track survey
}
\source{
Wildlife Conservation Society
}
\references{
Lubis et al., 2023
}
\examples{
data(track)
## maybe str(track) ; plot(track) ...
}
\keyword{datasets}
