% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track2dm_mean.R
\name{track2dm_mean}
\alias{track2dm_mean}
\title{Calculate mean from data.}
\usage{
track2dm_mean(myVector)
}
\arguments{
\item{myVector}{A vector contains numeric values.}
}
\value{
Mean of the vector.
}
\description{
A function to calculate mean from numeric vector.
}
