% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track2pts.R
\name{track2pts}
\alias{track2pts}
\title{Combine track and waypoints based on two common columns}
\usage{
track2pts(
  trackSp,
  track_id_1,
  track_id_2,
  minDist,
  waypointSp,
  point_id_1,
  point_id_2
)
}
\arguments{
\item{trackSp}{Digitised tracks in spatial dataframe.}

\item{track_id_1}{The first common column in tracks data (e.g., Patrol_ID).}

\item{track_id_2}{The second common column in tracks data (e.g., Patrol_Date).}

\item{minDist}{Minimum distance between points.}

\item{waypointSp}{Waypoints in spatial dataframe.}

\item{point_id_1}{The first common column in waypoints data (e.g., Patrol_ID).}

\item{point_id_2}{The second common column in waypoints data (e.g., Waypoint_Date).}
}
\value{
Spatial points dataframe.
}
\description{
A function that combine track and waypoints based on two common columns.
}
