% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2sp.R
\name{df2sp}
\alias{df2sp}
\title{df2sp}
\usage{
df2sp(dataFrame, UTMZone)
}
\arguments{
\item{dataFrame}{A data-frame contains at least "X" and "Y" and other columns to be converted to spatial object dataframe.}

\item{UTMZone}{A UTM projection of the target.}
}
\value{
Spatial object dataframe.
}
\description{
A function to convert data-frame to spatial object data-frame.
}
