% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist3D.R
\name{dist3D}
\alias{dist3D}
\title{dist3D}
\usage{
dist3D(dataFrame, elevData, repLength)
}
\arguments{
\item{dataFrame}{A dataframe contains at least "X", "Y" and "Z" in order to calculate distance using Eulidean distance formula.}

\item{elevData}{A raster layer contains elevation data in meter to calculate altitude (Z).}

\item{repLength}{An information about a desired length of each spatial replicate.}
}
\value{
New columns that specify distance and what replicates that the observations are belong to.
}
\description{
A function to calculate 3D distance and then categorise the observations based on a predefine replication length.
}
