% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist3D.R
\name{dist3D}
\alias{dist3D}
\title{Calculate distance in 3-dimension (X,Y,Z)}
\usage{
dist3D(dataFrame, Xcol, Ycol, elevData, repLength)
}
\arguments{
\item{dataFrame}{A dataframe contains pairs of coordinate in UTM projection to calculate distance using Euclidean distance formula.}

\item{Xcol}{A quoted name of column that consists X coordinates.}

\item{Ycol}{A quoted name of column that consists Y coordinates.}

\item{elevData}{A raster layer contains elevation data in meter to calculate altitude (Z).}

\item{repLength}{An information about a desired length of each spatial replicate.}
}
\value{
New columns that specify distance and what replicates that the observations are belong to.
}
\description{
A function to calculate 3D distance and then categories the observations based on a predefine replication length.
}
