% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm2spatial.R
\name{dm2spatial}
\alias{dm2spatial}
\title{Convert detection matrix into sf object}
\usage{
dm2spatial(detectMatrix, spProject)
}
\arguments{
\item{detectMatrix}{A matrix contains species detection/non-detection.}

\item{spProject}{spatial object where its projection to be copied.}
}
\value{
An sf data-frame contains species detection/non-detection.
}
\description{
to convert detection matrix to sf object.
}
