% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line2points.R
\name{line2points}
\alias{line2points}
\title{line2points}
\usage{
line2points(spLineDF, minDist)
}
\arguments{
\item{spLineDF}{A A SpatialLinesDataFrame of a track from a transect survey.}

\item{minDist}{Minimum distance between points.}
}
\value{
A SpatialPointsDataFrame from the line track.
}
\description{
A function to turn a line track into an ordered points along that line track
}
