% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeGrids.R
\name{makeGrids}
\alias{makeGrids}
\title{Create grid cells (fishnet) from a given spatial data}
\usage{
makeGrids(spObject, cellSize, clip = FALSE)
}
\arguments{
\item{spObject}{A spatial object in UTM projection.}

\item{cellSize}{Cell size for each subgrid in meters.}

\item{clip}{Clip grids within the spObject (default FALSE or preserve the whole grids).}
}
\value{
A SpatialPolygoneDataframe of grid cells.
}
\description{
A function to create fishnet from a given spatial object/extent.
}
