% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearID.R
\name{nearID}
\alias{nearID}
\title{nearID}
\usage{
nearID(wayPoints, trackPoints)
}
\arguments{
\item{wayPoints}{Spatial points dataframe of waypoints.}

\item{trackPoints}{Spatial points dataframe of track points (usually generated from GPX).}
}
\value{
Similar data with wayPoints with additional Id copied from nearby track points.
}
\description{
A function to copy IDs from other nearby track points.
}
