% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliceGrids.R
\name{sliceGrids}
\alias{sliceGrids}
\title{Split grid cells into smaller grid cells}
\usage{
sliceGrids(mainGrids, mainID, aggreFact, nRandom)
}
\arguments{
\item{mainGrids}{A SpatialPolygoneDataframe of the main gridcells to be sliced.}

\item{mainID}{A column in the main gridcells contain the cells ID.}

\item{aggreFact}{Aggregation factor expressed as number of cells in each direction (horizontally and vertically).}

\item{nRandom}{if nRandom is not FALSE, then randomize the subgrids.}
}
\value{
A spatialpolygonedataframe of smaller grid cells.
}
\description{
A function to create fishnet from a given spatial object/extent.
}
