% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speciesDM.R
\name{speciesDM}
\alias{speciesDM}
\title{speciesDM}
\usage{
speciesDM(
  speciesDF,
  sortID,
  Xcol,
  Ycol,
  whichCol,
  whichSp,
  samplingCov = FALSE,
  samplingFun = FALSE
)
}
\arguments{
\item{speciesDF}{A matrix contains Replicate column resulted from swts::dist3D() function.}

\item{sortID}{A name of column that used to order the point based on a sequence (time based or ID based).}

\item{Xcol}{A quoted name of column that consists X coordinates.}

\item{Ycol}{A quoted name of column that consists Y coordinates.}

\item{whichCol}{A column that contains all the species occurrence.}

\item{whichSp}{A selected species name within the "whichCol" column to be extracted.}

\item{samplingCov}{Sampling covariates to be extracted (default is FALSE)from each replicate.}

\item{samplingFun}{A list of method to deal with samplingCov (only modal (for character), mean (for numeric), and canopy (to calculate canopy closure) functions that are currently available.}
}
\value{
A data-frame contains detection matrix for selected species along with its geographic coordinates and sampling covariates.
}
\description{
A function to create detection matrix for selected species.
}
