% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeFormat.R
\name{timeFormat}
\alias{timeFormat}
\title{timeFormat}
\usage{
timeFormat(myString)
}
\arguments{
\item{myString}{Information about date and time from tracklogs.}
}
\value{
a new string with DateTime format (as.Date or as.POSIXct).


# Reformat the datetime
swts::trackTime(TrackSP = df, TimeZone = "UTC")
}
\description{
A function to reformat date and time.
}
